pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-os"
version = "1.0.0"
description = "A comprehensive product development workflow management system for AI coding platforms"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Agent OS Team", email = "agent-os@example.com"},
]
maintainers = [
    {name = "Agent OS Team", email = "agent-os@example.com"},
]
keywords = ["ai", "development", "workflow", "productivity", "coding", "agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "pyyaml>=5.4.0",
    "click>=8.0.0",
    "rich>=10.0.0",
    "pathlib2>=2.3.0; python_version < '3.4'",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
]
test = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
]

[project.urls]
Homepage = "https://github.com/fenghaitao/agent-os"
Repository = "https://github.com/fenghaitao/agent-os"
Documentation = "https://github.com/fenghaitao/agent-os#readme"
"Bug Tracker" = "https://github.com/fenghaitao/agent-os/issues"

[project.scripts]
agent-os = "agent_os.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_os*"]

[tool.setuptools.package-data]
agent_os = [
    "instructions/**/*",
    "standards/**/*", 
    "claude-code/**/*",
    "github-copilot/**/*",
    "qwen-code/**/*",
    "commands/**/*",
    "*.yml",
    "*.yaml",
]

[tool.setuptools.data-files]
"" = [
    "instructions/**/*",
    "standards/**/*",
    "claude-code/**/*", 
    "github-copilot/**/*",
    "qwen-code/**/*",
    "commands/**/*",
    "*.yml",
    "*.yaml",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
